Switch to voyager navigation + screen model
Switch to voyager navigation + screen model including a larger restructuring of the UI components.
