+++ desktop-app/src/main/kotlin/info/marozzo/hematoma/components/CompetitorsList.kt
@@ -1,0 +1,51 @@
+/*
+ * Copyright Jakob Ecker, 2025
+ * Licensed under the EUPL-1.2-or-later
+ *
+ */
+
+package info.marozzo.hematoma.components
+
+import androidx.compose.foundation.VerticalScrollbar
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.foundation.rememberScrollbarAdapter
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.ListItem
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import info.marozzo.hematoma.domain.Competitor
+import info.marozzo.hematoma.domain.CompetitorId
+import kotlinx.collections.immutable.ImmutableMap
+import kotlinx.collections.immutable.toImmutableList
+
+@Composable
+fun CompetitorsList(competitors: ImmutableMap<CompetitorId, Competitor>, modifier: Modifier = Modifier) {
+    val state = rememberLazyListState()
+    Box(modifier) {
+        LazyColumn(state = state) {
+            items(competitors.keys.toImmutableList(), key = { it }) {
+                CompetitorListItem(competitors[it]!!, modifier = Modifier.fillMaxWidth())
+            }
+        }
+        VerticalScrollbar(
+            modifier = Modifier.align(Alignment.CenterEnd).fillMaxHeight(),
+            adapter = rememberScrollbarAdapter(state)
+        )
+    }
+}
+
+@Composable
+private fun CompetitorListItem(competitor: Competitor, modifier: Modifier = Modifier) = Column(modifier) {
+    ListItem(
+        leadingContent = { Text("${competitor.registration.value}.") },
+        headlineContent = { Text(competitor.name.value) })
+    HorizontalDivider()
+}
